requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Deprecated==1.2.12
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.6.0
django-extensions==3.1.3
django-summernote==0.8.11.6
djangorestframework==3.12.4
drf-jwt==1.19.1
drf-yasg==1.20.0
environ==1.0
flex==6.14.1
idna==3.2
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
jsonpointer==2.1
MarkupSafe==2.0.1
numpy==1.21.1
packaging==21.0
pandas==1.3.1
Pillow==8.3.2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
rfc3987==1.3.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
strict-rfc3339==0.7
uritemplate==3.0.1
urllib3==1.26.6
validate-email==1.3
wrapt==1.12.1
